integrate the device side code into the Qt DevEmu GUI development helper

+++ device_side/src/Ui/Components/BatteryIndicator.cpp
@@ -17,5 +17,5 @@ void BatteryIndicator::draw(DisplayPtr display) {
     display->fillRect(mPosX+1, mPosY+5, (mWidth-2) * (batteryLevel / 100.0), mHeight, color_utils::WHITE);
     display->fillRect(mPosX+mWidth, mPosY+7, 2, 4, color_utils::WHITE);
     display->setCursor(mPosX + mWidth+6, mPosY + 2);
-    display->print(batteryLevel);
+    display->print(static_cast<int8_t>(batteryLevel));
 }
